requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-google-genai>=0.0.11
faiss-cpu>=1.7.4
python-dotenv>=1.0.1
PyPDF2>=3.0.1
python-docx>=1.1.0
pandas>=2.2.1
openpyxl>=3.1.2
python-pptx>=0.6.23 